backend/requirements.txt
requests==2.32.0
psycopg[binary]==3.1.18
Django
djangorestframework
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==22.1
python-dotenv==0.19.0
django-cors-headers
djoser==2.3.0
django-colorfield==0.9.0
gunicorn
celery[redis]==5.2.7
flower==1.2.0
celery_singleton==0.3.1
django-redis==5.2.0